Huge refactorization
