+++ tests/ui/test_templating.py
